rename to error_rule
